+++ src/pages/Home.tsx
@@ -3,59 +3,18 @@
 import SearchBar from "@/components/SearchBar";
 import SubwayState from "@/components/SubwayState";
 import { Button } from "@/components/ui/button";
+import useInstallPWA from "@/hooks/useInstallPWA";
 import useSubscriptionStatus from "@/hooks/useSubscriptionStatus";
 import useSearchResultStore from "@/stores/searchResult";
-import { useEffect, useState } from "react";
-
-interface BeforeInstallPromptEvent extends Event {
-  prompt: () => Promise<void>;
-  userChoice: Promise<{ outcome: "accepted" | "dismissed"; platform: string }>;
-}
-
-let defferedPrompt: BeforeInstallPromptEvent | null = null;
+import { useEffect } from "react";
 
 export default function Home() {
-  const [isInstallable, setIsInstallable] = useState(false);
   const setUserSubscriptionInfo = useSearchResultStore(
     (state) => state.setUserSubscriptionInfo
   );
+  const { isInstallable, showInstallPrompt } = useInstallPWA();
 
   const { CheckSubscriptionStatus } = useSubscriptionStatus();
-
-  useEffect(() => {
-    const handleBeforeInstallPrompt = (e: BeforeInstallPromptEvent) => {
-      e.preventDefault();
-      defferedPrompt = e;
-      setIsInstallable(true);
-    };
-
-    window.addEventListener(
-      "beforeinstallprompt",
-      handleBeforeInstallPrompt as EventListener
-    );
-
-    return () => {
-      window.removeEventListener(
-        "beforeinstallprompt",
-        handleBeforeInstallPrompt as EventListener
-      );
-    };
-  }, []);
-
-  const showPWAInstallPrompt = () => {
-    if (defferedPrompt) {
-      defferedPrompt.prompt();
-      defferedPrompt.userChoice.then((choiceResult) => {
-        if (choiceResult.outcome === "accepted") {
-          console.log("사용자가 설치를 수락함");
-        } else {
-          console.log("사용자가 설치를 거절함");
-        }
-        defferedPrompt = null;
-        setIsInstallable(false);
-      });
-    }
-  };
 
   const unSubscribeUser = async (subscription: PushSubscription) => {
     const unSubscribeStatus = await subscription.unsubscribe();
@@ -156,7 +115,7 @@ export default function Home() {
   return (
     <div id="home" className="flex flex-col gap-16 max-w-[418px] w-full mt-4">
       {isInstallable && (
-        <Button onClick={showPWAInstallPrompt}>PWA를 설치하세용</Button>
+        <Button onClick={showInstallPrompt}>PWA를 설치하세용</Button>
       )}
       <Button onClick={requestPushPermission}>푸시 알림 허용</Button>
       <SearchBar />
